+++ JsonToolsNppPlugin/JSONTools/JsonGrepper.cs
@@ -3,11 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading;
-using System.Threading.Tasks;
 using System.Net;
-using System.Text;
-using System.Windows.Forms;
-using JSON_Tools.Utils;
 /*
 using System;
 using System.Net;