+++ JsonToolsNppPlugin/Tests/YamlDumperTests.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using JSON_Tools.JSON_Tools;
+﻿using JSON_Tools.JSON_Tools;
 using JSON_Tools.Utils;
-using Kbg.NppPluginNET.PluginInfrastructure;
 
 namespace JSON_Tools.Tests
 {